+++ pyproject.toml
@@ -21,7 +21,7 @@ license = {text = "Apache-2.0"}
 # ONLY dependencies that are actually imported/used in the code
 dependencies = [
     # Core MCP framework - imported in src/prompts/__init__.py and src/core/server.py
-    "fastmcp==2.10.3",
+    "fastmcp==2.11.3",
     
     # Instana API client - imported in src/event/events_tools.py
     "instana_client==1.0.0",
