+++ src/settings.py
@@ -1,0 +1,8 @@
+## Global variables
+
+# Instana backend API related credentials
+global_instana_api_token = None
+global_instana_base_url = None
+
+# List of tool categories to enable, e.g., ["infra", "app", "events", "automation", "website", "log"]
+global_tool_categories = None
