+++ src/tools/__init__.py
@@ -1,0 +1,1 @@
+from tools.log import demo_tool as log_demo_tools
