+++ src/tools/application/application_global_alert_config.py
