+++ src/tools/infrastructure/infrastructure_analyze.py
